% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtkore.R
\docType{package}
\name{rtkore}
\alias{rtkore}
\alias{rtkore-package}
\title{rtkore is a R/STK++ bridge}
\description{
This package contains the header files for STK++ core library.
The typical usage is to install this package and list it in
the \env{LinkingTo: } line in the \file{DESCRIPTION} file of
other packages.
}
\details{
As described at the STK++ project's home page, \url{http://www.stkpp.org},
STK++ is a versatile, fast, reliable and elegant collection of C++ classes
for statistics, clustering, classification, regression, dimension reduction,
linear algebra, etc.

The core package provide a set arrays (with an Eigen-like API) allowing
to store and process data in an efficient way.

\tabular{ll}{
  Package: \tab rtkore\cr
  Type: \tab Package\cr
  Version: \tab 0.8.1\cr
  Date: \tab 2014-07-05\cr
  License: \tab GPL for the rtkore side, LGPL for the stkpp side  + file LICENSE\cr
  LazyLoad: \tab yes\cr
}
}
\author{
Author: Serge Iovleff \email{contact@stkpp.org}
}
\keyword{STK++,}
\keyword{stkpp}

