% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtkoreFlags.R
\name{CxxFlags}
\alias{CxxFlags}
\alias{CppFlags}
\alias{LdFlags}
\title{CxxFlags defaults for rtkore package}
\usage{
CxxFlags(cpp11 = FALSE)

CppFlags()

LdFlags()
}
\description{
CxxFlags defaults for rtkore package

CppFlags defaults for rtkore package

LdFlags defaults for rtkore package
}
\keyword{internal}
