% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtkoreFlags.R
\name{CxxFlags}
\alias{CppFlags}
\alias{CxxFlags}
\alias{LdFlags}
\title{CxxFlags defaults for the rtkore}
\usage{
CxxFlags(cpp11 = FALSE)

CppFlags()

LdFlags()
}
\description{
CxxFlags defaults for the rtkore

CppFlags defaults for the rtkore

LdFlags defaults for the rtkore package
}
\keyword{internal}

